Refatoração
Foi realizada uma refatoração no código para seguir boas práticas de programação.
