+++ engenharia_disciplinada_de_software/src/app/Componentes/BarraEsquerda/MenuItemWithIcon.tsx
@@ -1,0 +1,27 @@
+// BarraEsquerda/MenuItemWithIcon.tsx
+"use client";
+
+import { MenuItem, IconWrapper, IconLabel } from "./styles";
+
+interface MenuItemWithIconProps {
+  icon: string;
+  label: string;
+  selected?: boolean;
+  expanded: boolean;
+  onClick: () => void;
+}
+
+export function MenuItemWithIcon({
+  icon,
+  label,
+  selected,
+  expanded,
+  onClick,
+}: MenuItemWithIconProps) {
+  return (
+    <MenuItem $selected={selected} onClick={onClick}>
+      <IconWrapper><i className={icon} /></IconWrapper>
+      {expanded && <IconLabel>{label}</IconLabel>}
+    </MenuItem>
+  );
+}
