+++ engenharia_disciplinada_de_software/src/app/Componentes/BarraEsquerda/page.tsx
@@ -1,21 +1,20 @@
-'use client';
+// BarraEsquerda/BarraEsquerda.tsx
+"use client";
 
-import { useState, useEffect } from "react";
+import { useState } from "react";
 import Image from "next/image";
+import { usePathname, useRouter } from "next/navigation";
+import { useUserData } from "./useUserData";
+import { MenuItemWithIcon } from "./MenuItemWithIcon";
+
 import {
   Container,
   MenuWrapper,
-  IconWrapper,
-  IconLabel,
   UserInfo,
   UserName,
   UserStatus,
-  MenuItem,
   Footer
 } from "./styles";
-
-import { usePathname, useRouter } from "next/navigation";
-import { BlogUser, fetchUserData } from "@/app/API/UserAPI/ApiUserData";
 
 import { Montserrat, Poppins } from "next/font/google";
 export const montserrat = Montserrat({ subsets: ["latin"], weight: ["400", "700"] });
@@ -23,21 +22,11 @@ export const poppins = Poppins({ subsets: ["latin"], weight: ["400", "700"] });
 
 export default function BarraEsquerda() {
   const [expanded, setExpanded] = useState(false);
-  const [user, setUser] = useState<BlogUser | null>(null);
-  const [token, setToken] = useState<string | null>(null);
+  const { user, token } = useUserData();
   const router = useRouter();
   const pathname = usePathname();
 
-  useEffect(() => {
-    const loadUser = async () => {
-      const storedToken = localStorage.getItem("token");
-      setToken(storedToken);
-
-      const data = await fetchUserData();
-      setUser(data);
-    };
-    loadUser();
-  }, []);
+  const handleGoTo = (path: string) => router.push(path);
 
   return (
     <div className={poppins.className}>
@@ -52,7 +41,7 @@ export default function BarraEsquerda() {
               <Image src="/Perfil.svg" alt="User" width={50} height={50} />
               <div>
                 <UserName>{user.name}</UserName>
-                {user && token ? (
+                {token ? (
                   <UserStatus>🟢 Ativo</UserStatus>
                 ) : (
                   <UserStatus>🔴 Offline</UserStatus>
@@ -61,36 +50,47 @@ export default function BarraEsquerda() {
             </UserInfo>
           )}
 
-          <MenuItem $selected={pathname === "/Home"} onClick={() => router.push("/Home")}>
-            <IconWrapper><i className="fas fa-home" /></IconWrapper>
-            {expanded && <IconLabel>Home</IconLabel>}
-          </MenuItem>
-
-          <MenuItem onClick={() => router.push("/Search")}>
-            <IconWrapper><i className="fas fa-search" /></IconWrapper>
-            {expanded && <IconLabel>Search</IconLabel>}
-          </MenuItem>
-
-          <MenuItem onClick={() => router.push("/Categorias")}>
-            <IconWrapper><i className="fas fa-layer-group" /></IconWrapper>
-            {expanded && <IconLabel>Categories</IconLabel>}
-          </MenuItem>
-
-          <MenuItem $selected={pathname === "/User"} onClick={() => router.push("/User")}>
-            <IconWrapper><i className="fas fa-user" /></IconWrapper>
-            {expanded && <IconLabel>User</IconLabel>}
-          </MenuItem>
+          <MenuItemWithIcon
+            icon="fas fa-home"
+            label="Home"
+            expanded={expanded}
+            selected={pathname === "/Home"}
+            onClick={() => handleGoTo("/Home")}
+          />
+          <MenuItemWithIcon
+            icon="fas fa-search"
+            label="Search"
+            expanded={expanded}
+            onClick={() => handleGoTo("/Search")}
+          />
+          <MenuItemWithIcon
+            icon="fas fa-layer-group"
+            label="Categories"
+            expanded={expanded}
+            onClick={() => handleGoTo("/Categorias")}
+          />
+          <MenuItemWithIcon
+            icon="fas fa-user"
+            label="User"
+            expanded={expanded}
+            selected={pathname === "/User"}
+            onClick={() => handleGoTo("/User")}
+          />
         </MenuWrapper>
 
         <Footer>
-          <MenuItem>
-            <IconWrapper><i className="fas fa-cog" /></IconWrapper>
-            {expanded && <IconLabel>Settings</IconLabel>}
-          </MenuItem>
-          <MenuItem>
-            <IconWrapper><i className="fas fa-exclamation-circle" /></IconWrapper>
-            {expanded && <IconLabel>Support</IconLabel>}
-          </MenuItem>
+          <MenuItemWithIcon
+            icon="fas fa-cog"
+            label="Settings"
+            expanded={expanded}
+            onClick={() => {}}
+          />
+          <MenuItemWithIcon
+            icon="fas fa-exclamation-circle"
+            label="Support"
+            expanded={expanded}
+            onClick={() => {}}
+          />
         </Footer>
       </Container>
     </div>
