+++ engenharia_disciplinada_de_software/src/app/Componentes/CriarPost/page.tsx
@@ -1,6 +1,6 @@
+// Componentes/CriarPost/page.tsx
 "use client";
 
-import { useEffect, useState } from "react";
 import {
   PostContainer,
   TextArea,
@@ -9,44 +9,21 @@ import {
   ActionButton,
   SendButton,
 } from "./styles";
-import { createPost } from "../../API/Posts/WritePost/WritePostApi";
-import { getAllCategories, CategoriaData } from "../../API/Categorias/PostCategorias";
 
 import { Montserrat, Poppins } from "next/font/google";
+import { useCriarPost } from "./useCriarPost";
+
 export const montserrat = Montserrat({ subsets: ["latin"], weight: ["400", "700"] });
 export const poppins = Poppins({ subsets: ["latin"], weight: ["400", "700"] });
 
 export default function CriarPost() {
-  const [title, setTitle] = useState("");
-  const [post, setPost] = useState("");
-  const [categorias, setCategorias] = useState<CategoriaData[]>([]);
-  const [selectedCategoria, setSelectedCategoria] = useState<number | null>(null);
-
-  useEffect(() => {
-    const fetchCategorias = async () => {
-      const data = await getAllCategories();
-      setCategorias(data);
-      if (data.length > 0) setSelectedCategoria(data[0].id);
-    };
-    fetchCategorias();
-  }, []);
-
-  const handleSend = async () => {
-    if (!selectedCategoria || post.trim() === "" || title.trim() === "") return;
-
-    const response = await createPost({
-      title,
-      content: post,
-      categoryId: selectedCategoria,
-    });
-
-    if (response) {
-      console.log("Post criado com sucesso:", response);
-      setPost("");
-      setTitle("");
-      setSelectedCategoria(categorias[0]?.id || null);
-    }
-  };
+  const {
+    title, setTitle,
+    post, setPost,
+    categorias,
+    selectedCategoria, setSelectedCategoria,
+    handleSend,
+  } = useCriarPost();
 
   return (
     <div className={poppins.className}>
@@ -65,7 +42,6 @@ export default function CriarPost() {
           onChange={(e) => setPost(e.target.value)}
         />
 
-        {/* Dropdown de categorias */}
         <select
           value={selectedCategoria ?? ""}
           onChange={(e) => setSelectedCategoria(Number(e.target.value))}
