+++ engenharia_disciplinada_de_software/src/app/Componentes/Posts/PostContainer.tsx
@@ -1,13 +1,5 @@
+// Posts/PostContainer.tsx
 "use client";
-
-import { useEffect, useState } from "react";
-import { getAllPosts, PostData } from "../../API/Posts/GetPosts/GetPostsAPI";
-import { deletePostById } from "../../API/Posts/DeletePost/DeletePost";
-import { updatePost } from "../../API/Posts/PutPost/EditarPost";
-import { postComment } from "../../API/Comments/PostComents/PostComentsAPI";
-import { getAllComments, CommentData } from "../../API/Comments/GetComents/GetComents";
-import { useSearchParams } from "next/navigation";
-import { getPostsByCategory } from "../../API/Posts/GetPostCategory/GetPostCategory";
 
 import {
   PostWrapper,
@@ -26,100 +18,18 @@ import {
 } from "./styles";
 
 import { Montserrat, Poppins } from "next/font/google";
+import { usePostContainer } from "./usePostContainer";
+
 export const montserrat = Montserrat({ subsets: ["latin"], weight: ["400", "700"] });
 export const poppins = Poppins({ subsets: ["latin"], weight: ["400", "700"] });
 
 export default function PostContainer() {
-  const [posts, setPosts] = useState<PostData[]>([]);
-  const [userId, setUserId] = useState<number | null>(null);
-  const [editingPostId, setEditingPostId] = useState<number | null>(null);
-  const [editedTitle, setEditedTitle] = useState("");
-  const [editedContent, setEditedContent] = useState("");
-  const [commentInput, setCommentInput] = useState<Record<number, string>>({});
-  const [comments, setComments] = useState<CommentData[]>([]);
-  const searchParams = useSearchParams();
-  const categoryName = searchParams.get("category");
-
-  useEffect(() => {
-    const fetchData = async () => {
-      let postsData = null;
-
-      if (categoryName) {
-        postsData = await getPostsByCategory(categoryName);
-      } else {
-        postsData = await getAllPosts();
-      }
-
-      const commentsData = await getAllComments();
-
-      if (postsData) setPosts(postsData);
-      setComments(commentsData);
-    };
-
-    const token = localStorage.getItem("token");
-    if (token) {
-      const decoded = parseJwt(token);
-      setUserId(decoded?.id || null);
-    }
-
-    fetchData();
-  }, [categoryName]);
-
-  const handleDelete = async (postId: number) => {
-    const confirmDelete = confirm("Deseja realmente excluir este post?");
-    if (!confirmDelete) return;
-
-    const success = await deletePostById(postId);
-    if (success) {
-      setPosts((prev) => prev.filter((p) => p.id !== postId));
-    }
-  };
-
-  const handleEdit = (post: PostData) => {
-    setEditingPostId(post.id);
-    setEditedTitle(post.title);
-    setEditedContent(post.content);
-  };
-
-  const handleSubmitEdit = async (postId: number) => {
-    const success = await updatePost({
-      id: postId,
-      title: editedTitle,
-      content: editedContent,
-      categoryId: 1,
-    });
-
-    if (success) {
-      setPosts((prev) =>
-        prev.map((p) =>
-          p.id === postId ? { ...p, title: editedTitle, content: editedContent } : p
-        )
-      );
-      setEditingPostId(null);
-    }
-  };
-
-  const handleCommentSubmit = async (postId: number) => {
-    const content = commentInput[postId]?.trim();
-    if (!content) return;
-
-    const success = await postComment({
-      content,
-      approved: true,
-      updatedAt: new Date().toISOString(),
-    });
-
-    if (success) {
-      alert("Comentário enviado!");
-      setCommentInput((prev) => ({ ...prev, [postId]: "" }));
-
-      // Atualiza a lista de comentários
-      const updatedComments = await getAllComments();
-      setComments(updatedComments);
-    } else {
-      alert("Erro ao enviar comentário.");
-    }
-  };
+  const {
+    posts, userId, editingPostId, editedTitle, editedContent,
+    commentInput, comments,
+    setEditedTitle, setEditedContent, setCommentInput,
+    handleDelete, handleEdit, handleSubmitEdit, handleCommentSubmit,
+  } = usePostContainer();
 
   return (
     <div className={poppins.className}>
@@ -195,22 +105,4 @@ export default function PostContainer() {
       ))}
     </div>
   );
-}
-
-function parseJwt(token: string) {
-  try {
-    const base64Url = token.split(".")[1];
-    const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
-    const jsonPayload = decodeURIComponent(
-      atob(base64)
-        .split("")
-        .map((c) => "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2))
-        .join("")
-    );
-
-    return JSON.parse(jsonPayload);
-  } catch (e) {
-    console.error("Erro ao decodificar token:", e);
-    return null;
-  }
 }
