+++ engenharia_disciplinada_de_software/src/app/Componentes/Posts/parseJwt.ts
@@ -1,0 +1,18 @@
+// Posts/parseJwt.ts
+export function parseJwt(token: string) {
+    try {
+      const base64Url = token.split(".")[1];
+      const base64 = base64Url.replace(/-/g, "+").replace(/_/g, "/");
+      const jsonPayload = decodeURIComponent(
+        atob(base64)
+          .split("")
+          .map((c) => "%" + ("00" + c.charCodeAt(0).toString(16)).slice(-2))
+          .join("")
+      );
+      return JSON.parse(jsonPayload);
+    } catch (e) {
+      console.error("Erro ao decodificar token:", e);
+      return null;
+    }
+  }
+
