+++ engenharia_disciplinada_de_software/src/app/Login/auth.ts
@@ -1,0 +1,9 @@
+// Login/auth.ts
+export function setAuthToken(token: string) {
+    localStorage.setItem("token", token);
+  }
+  
+  export function getAuthToken(): string | null {
+    return localStorage.getItem("token");
+  }
+
