+++ engenharia_disciplinada_de_software/src/app/Login/page.tsx
@@ -1,12 +1,9 @@
+// Login/Login.tsx
 "use client";
 
 import Link from "next/link";
 import Image from "next/image";
-import { useState } from "react";
-import { useRouter } from "next/navigation";
 import "@fortawesome/fontawesome-free/css/all.min.css";
-import { loginUsuario } from "./apiServiceLogin";
-
 import {
   Container,
   LeftSection,
@@ -31,66 +28,19 @@ import {
   LoginImage,
 } from "./styles";
 
+import { useLogin } from "./useLogin";
+
 export default function Login() {
-  const router = useRouter();
-
-  const [form, setForm] = useState({
-    email: "",
-    password: "",
-  });
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setForm({ ...form, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = async () => {
-    if (!form.email || !form.password) {
-      alert("Preencha todos os campos.");
-      return;
-    }
-
-    try {
-      const result = await loginUsuario({
-        email: form.email,
-        password: form.password,
-      });
-
-      console.log("Resultado do login:", result); 
-
-      if (result.jwt) {
-        localStorage.setItem("token", result.jwt);
-        router.push("/Home");
-        alert("Login realizado com sucesso!");
-      } else {
-        throw new Error("Token ou ID não recebido do servidor.");
-      }
-    } catch (error: any) {
-      alert("Erro: " + error.message);
-    }
-  };
+  const { form, handleChange, handleSubmit } = useLogin();
 
   return (
-    
     <Container>
       <LeftSection>
         <LoginImage>
-          <Image
-            src="/Login.svg"
-            alt="Login Illustration"
-            width={600}
-            height={600}
-            style={{ width: "100%", height: "auto" }}
-            priority
-          />
+          <Image src="/Login.svg" alt="Login Illustration" width={600} height={600} style={{ width: "100%", height: "auto" }} priority />
         </LoginImage>
         <LogoContainer>
-          <Image
-            src="/Logo.svg"
-            alt="Page Flow Logo"
-            width={600}
-            height={50}
-            style={{ width: "100%", height: "auto" }}
-          />
+          <Image src="/Logo.svg" alt="Page Flow Logo" width={600} height={50} style={{ width: "100%", height: "auto" }} />
         </LogoContainer>
         <Phrase>Escreva. Compartilhe. Conecte-se.</Phrase>
       </LeftSection>
@@ -128,13 +78,7 @@ export default function Login() {
           <label className="text-xl mt-4">Password</label>
           <InputContainer>
             <InputIcon>🔒</InputIcon>
-            <InputField
-              type="password"
-              name="password"
-              placeholder="Password ..."
-              value={form.password}
-              onChange={handleChange}
-            />
+            <InputField type="password" name="password" placeholder="Password ..." value={form.password} onChange={handleChange} />
           </InputContainer>
 
           <SignInButton onClick={handleSubmit}>Sign In</SignInButton>
