+++ engenharia_disciplinada_de_software/src/app/Login/useLogin.ts
@@ -1,0 +1,43 @@
+// Login/useLogin.ts
+"use client";
+
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import { loginUsuario } from "./apiServiceLogin";
+import { setAuthToken } from "./auth";
+
+export function useLogin() {
+  const router = useRouter();
+  const [form, setForm] = useState({ email: "", password: "" });
+
+  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setForm({ ...form, [e.target.name]: e.target.value });
+  };
+
+  const handleSubmit = async () => {
+    if (!form.email || !form.password) {
+      alert("Preencha todos os campos.");
+      return;
+    }
+
+    try {
+      const result = await loginUsuario(form);
+
+      if (result.jwt) {
+        setAuthToken(result.jwt);
+        router.push("/Home");
+        alert("Login realizado com sucesso!");
+      } else {
+        throw new Error("Token ou ID não recebido do servidor.");
+      }
+    } catch (error: any) {
+      alert("Erro: " + error.message);
+    }
+  };
+
+  return {
+    form,
+    handleChange,
+    handleSubmit,
+  };
+}
